Failing extractors because unable to pull-off type from reflection
